31 refactor the save logic to be more streamlined (#38)
* Created the method and Lists for elements to save

Created the method and Lists for elements to save.

* Changed all of the loading of the Web Applications

Changed all of the loading of the Web Applications to use the generic LoadWebApplicationConfiguration() method.

* Converted loading/saving to generic interface class

Converted loading/saving to generic interface class. By doing so, it allowed me to simplify the redundant code that is out there and to simply utilize centralized data structure objects to tell me what elements and all need to be saved.

Also updated all of the automated tests, especially for the Web Application Data Structures tests as I added a bunch of new List<string> objects for the hiding of elements as well as the saving of the elements.