Added duplicate Meeting Type Name checking
Added code to check whether there are duplicate meeting types or not in the currently selected object as new Meeting Type Names values are added or updated.

Just need to add something to the main Integration element that gives a visual that there are duplicate Meeting Type Name values in a specific Integration, especially upon initial loading of the configuration.