2 create a dark mode (#44)
* Created the initial Dark Mode Code and Styling

Created the initial Dark Mode Code and Styling. Just need to work on making a way to retain the previous choice of the users session.

* Pushing more work up

* Finished adding code for dark mode

Finished adding code for dark mode. Can't find any other ones that need it all added, but if someone finds them, then a new Issue will be opened.