Added dynamic tooltips (#45)
Added dynamic tooltips for the Agenda Online web application. So far, everything appears to be working good. I know that there are still tooltips to add (as in the descriptions), but I will get to the blank ones as soon as I can get the rest of the Web Applications tooltips up and working.