Added Patient Window dynamic Tooltips
Added dynamic tooltips for the Patient Window web application. So far, everything appears to be working good. I know that there are still tooltips to add (as in the descriptions), but I will get to the blank ones as soon as I can get the rest of the Web Applications tooltips up and working.

Also updated all of the test suites and got them all back to passing after adding the new updates to the data structure.