Added Public Access - Legacy dynamic Tooltips (#53)
Added dynamic tooltips for the Public Access - Legacy web application. So far, everything appears to be working good. I know that there are still tooltips to add (as in the descriptions), but I will get to the blank ones as soon as I can get the rest of the Web Applications tooltips up and working.