Multiple Fixes/Enhancements
Enhanced by combining the error arrays into a single array.

Enhanced by moving a method that will detect duplicate objects in an array into it's own separate method to use for other items where I need to detect duplicate items.

Fixed the dark mode so that when you load the tables, the every other row is colored appropriately and the table header row is darker as well.

Also fixed an alert modal that I missed adding the dark_mode class to when the switch is toggled.