40 add the ability to create a new web application from scratch (#82)
* Created all of the core look and functionality

Created all of the core look and functionality that is needed for the modal to create a new configuration from scratch. Right now, I just have the Web Application button visible, but eventually I will add the Client Configuration button visible.

* Created the API endpoint for GetNewConfigurationDetails

Created the API endpoint for GetNewConfigurationDetails

* Created NewConfigurationDetails parsing logic

Created NewConfigurationDetails parsing logic.

* Added JavaScript code to parse version and Web Applications

Added JavaScript code to parse version and Web Applications and to then create the appropriate select optgroup and option elements.

* Created all of the UI elements and logic

Created all of the UI elements and logic to be able to create a new Web Application from scratch. I also added some logic and elements to handle future scenarios to make it all easier to add down the line.

* Fixed some issues and added more functionality

Fixed a couple of JavaScript logic issues as well as fixed the code to pull the VirtualDirectories into the object.

Finished the code to start the creation of the Web Application and create the root Endpoint URL.

* Put in duplicate WebApp/AppPool check

Put in duplicate WebApp/AppPool check to validate that the folder doesn't already exist, the Application doesn't already exist, or the Application Pool doesn't already exist.

* Created the code to create the new Web Application

Created the code to create the new Web Application and added appropriate try/catches.

* Completed the code to create web app from scratch

Completed the code to create web app from scratch.

* Added Dark Mode styling

Added Dark Mode styling to all of the appropriate items.