Development v0.9.0 (#83)
* Finished Additional Options Configuration

Finished the code to process the Additional Options text that is needed. Will work on getting the "Testing" button up and running when I create the branch for Issue #25 .

* Added the ability to test Connection Strings

Added the ability to test Connection Strings and shows the appropriate result value in the UI.

* Multiple Fixes/Enhancements

Enhanced by combining the error arrays into a single array.

Enhanced by moving a method that will detect duplicate objects in an array into it's own separate method to use for other items where I need to detect duplicate items.

Fixed the dark mode so that when you load the tables, the every other row is colored appropriately and the table header row is darker as well.

Also fixed an alert modal that I missed adding the dark_mode class to when the switch is toggled.

* Added check that Cache Path wasn't a UNC path

Added check that Cache Path wasn't a UNC path

* Updated Responsive Apps App display and backend

Updated Responsive Apps App display and backend. Also updated one of the automated tests to account for an App being in the web.config file.

* Updated the section outlined in the title

Updated the section for the FormattedTextIframeSupportedDomains so that it had it's own section and all. Also updated all of the automated tests and more.

* Validated an error is queued

After setting a configuration value above the maximum, it did in fact work and queued up an error in the error notification bell.

Also fixed the flashing of the box when  clicking the view button as the timing wasn't right.

* Completed duplicate logic revamp

Completed duplicate logic revamp on the Connection Strings section.

* 40 add the ability to create a new web application from scratch (#82)

* Created all of the core look and functionality

Created all of the core look and functionality that is needed for the modal to create a new configuration from scratch. Right now, I just have the Web Application button visible, but eventually I will add the Client Configuration button visible.

* Created the API endpoint for GetNewConfigurationDetails

Created the API endpoint for GetNewConfigurationDetails

* Created NewConfigurationDetails parsing logic

Created NewConfigurationDetails parsing logic.

* Added JavaScript code to parse version and Web Applications

Added JavaScript code to parse version and Web Applications and to then create the appropriate select optgroup and option elements.

* Created all of the UI elements and logic

Created all of the UI elements and logic to be able to create a new Web Application from scratch. I also added some logic and elements to handle future scenarios to make it all easier to add down the line.

* Fixed some issues and added more functionality

Fixed a couple of JavaScript logic issues as well as fixed the code to pull the VirtualDirectories into the object.

Finished the code to start the creation of the Web Application and create the root Endpoint URL.

* Put in duplicate WebApp/AppPool check

Put in duplicate WebApp/AppPool check to validate that the folder doesn't already exist, the Application doesn't already exist, or the Application Pool doesn't already exist.

* Created the code to create the new Web Application

Created the code to create the new Web Application and added appropriate try/catches.

* Completed the code to create web app from scratch

Completed the code to create web app from scratch.

* Added Dark Mode styling

Added Dark Mode styling to all of the appropriate items.

* Cleaned up old .zip files