Development v1.0.0 (#86)
This PR adds in one of the biggest and best additions to the COMACON utility...AUTHENTICATION! This release adds the following functionality:

1. Authentication
2. Password reset (Forgot Password functionality will be added in a future release)
3. User Management
4. Audit History (not visually showing yet, but hopefully soon enough)

And I am so happy that I feel comfortable enough to be able to release this out as the official V1.0.0 release. I can't wait to hear back from others about this and your ideas on how I can make it better.

Thank you to everyone who has supported me in this endeavor so far and I can't wait to see what the future holds for this utility!